Move ProxyFactory, WindowsProxy to app, simplify
