+++ app/ProxyFactory.h
@@ -15,27 +15,16 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-#include "core/ProxyFactory.h"
+#pragma once
 
 #include "core/Proxy.h"
 
-#if defined(Q_OS_WIN)
-#include "win/WindowsProxy.h"
-#endif
+#include <QObject>
 
-using namespace ama;
+class ProxyFactory
+{   
+public:
+    ProxyFactory();
 
-ProxyFactory::ProxyFactory()
-{
-
-}
-
-Proxy* ProxyFactory::create(const int port, QObject* parent)
-{
-#if defined(Q_OS_WIN)
-    return new ama::win::WindowsProxy(port, parent);
-#else
-#warning No platform support implemented for this OS, returning generic proxy.
-    return new Proxy(port, parent);
-#endif
-}
+    ama::Proxy* create(const int port, QObject* parent = nullptr);
+};
